Fixed a warning.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@27382 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/cmdline.c
@@ -197,7 +197,7 @@ int cmdline_parse(int *argc, char **argv)
                     break;
                 }
                 /* This is a kludge to allow --long options */
-                for (j = 0; j < strlen(argv[i]); j++) {
+                for (j = 0; j < (int)strlen(argv[i]); j++) {
                     argv[i][j] = argv[i][j + 1];
                 }
             }
